remove backward_compatibility folder
