+++ tests/integration/connection_strategy_test.py
