+++ tests/integration/lifecycle_test.py
