+++ tests/integration/predicate_test.py
