+++ tests/integration/proxy/cp/atomic_reference_test.py
@@ -1,6 +1,6 @@
 from hazelcast.errors import DistributedObjectDestroyedError, ClassCastError
 from hazelcast.serialization.api import IdentifiedDataSerializable
-from tests.integration.backward_compatibility.proxy.cp import CPTestCase
+from tests.integration.proxy.cp import CPTestCase
 from tests.util import set_attr
