+++ tests/integration/proxy/cp/hazelcast_cpsubsystem.xml
