+++ tests/integration/proxy/distributed_objects_test.py
