+++ tests/integration/proxy/hazelcast.xml
