+++ tests/integration/proxy/hazelcast_litemember.xml
