+++ tests/integration/proxy/hazelcast_mapstore.xml
