+++ tests/integration/proxy/map_test.py
