+++ tests/integration/proxy/ringbuffer_test.py
