+++ tests/integration/ssl_tests/client1-cert.pem
